Introduce BaseController class
